move structure validation to BaseContent from ContentItem. edit pydantic error message
