+++ demisto_sdk/commands/content_graph/strict_objects/base_strict_model.py
@@ -105,6 +105,14 @@ class StructureError(BaseStrictModel):
     ctx: Optional[dict] = None
 
     def __str__(self):
+        if self.error_type == "assertion_error":
+            self.error_message = f"The field {self.field_name} is not required, but should not be None if it exists"
+        elif self.error_type == "value_error.extra":
+            self.error_message = (
+                f"The field {self.field_name} is extra and {self.error_message}"
+            )
+        elif self.error_type == "value_error.missing":
+            self.error_message = f"The field {self.field_name} is required but missing"
         return (
             f"problematic field: {self.field_name} | error message: {self.error_message} |"
             f" error type : {self.error_type}"
